Add exception handling module
After some trial and error, a simple exception handling mechanism based
on `setjmp()` and `longjmp()` has been developed. The exception handling
module provides the following interface:
  - type `pcr_exception`
  - type `PCR_EXCEPTION`
  - constant `PCR_EXCEPTION_NONE`
  - constant `PCR_EXCEPTION_HANDLE`
  - constant `PCR_EXCEPTION_STATE`
  - constant `PCR_EXCEPTION_RANGE`
  - macro `pcr_exception_try()`
  - macro `pcr_exception_catch()`
  - macro `pcr_exception_catchall`
  - macro `pcr_exception_throw()`
  - macro `pcr_exception_trace()`
  - macro `pcr_exception_unwind()`

The interface for the exception handling module has been tested in the
`test/runner.c` file. As of now, this is a very rough test that is meant
to be verified by eye. A very simple Makefile has been introduced to
build and execute the test runner; this Makefile is expected to grow in
detail and complexity as the PCR Library develops.

In addition to the exception handling interface, the following set of
compiler hints have also been defined:
  - `pcr_hint_hot`
  - `pcr_hint_cold`
  - `pcr_hint_likely()`
  - `pcr_hint_unlikely()`

These compiler hint macros are available only GCC-compatible compilers,
which is what we would currently be restricting ourselves to.

Signed-off-by: Abhishek Chakravarti <abhishek@taranjali.org>