Add string module
The string module has been implemented in a very simplistic manner with
a minimal interface. Strings are handled by the `pcr_string` type, which
supports UTF-8 and is compatible with ordinary `char *` C strings. The
following interface is provided by this module:
  - pcr_string_new()
  - pcr_string_copy()
  - pcr_string_len()
  - pcr_string_add()

For now, this minimal API will suffice, unless string manipulation is
required. I initially thought of adding some string manipulation
routines, but it quickly became apparent that it is difficult to
implement indexing owing to the variable length of UTF-8 codepoints.
I'll need to think a bit more on this.

In the meantime, Sheredom's UTF-8 library has been used in the
implementation of the `pcr_string` interface.

The unit testing module has been refactored with `pcr_string`, thereby
providing a rough test of the string module. So far, everything seems
OK, although some Unicode chars are not being displayed correctly,
although I think that's more to do with my `urxvt` setup than anything
else.

Signed-off-by: Abhishek Chakravarti <abhishek@taranjali.org>