Add incomplete test cases for SQLite support
When I started out this branch, I had originally intended to complete
the test cases for all the interfaces that led to the `pcr_dbase`
interface running on the SQLite adapter.

In retrospect, this may have been biting off a bit too much at one go,
and I should have perhaps continued on an interface-by-interface
approach. Although many test cases have been implemented, by no means is
the coverage complete.

However, I am deferring the testing in favour of refactoring and
documenting the existing interfaces. It has become apparent that at this
stage, it is important to have some documentation in the form of
comments. Subsequently, I plan to follow through with improving and
thoroghly testing each interface, with one branch dedicated to each
interface.

As of now, in addition to the test cases that have already been
introduced, several enhancements to the API have been made where
relevant.

Signed-off-by: Abhishek Chakravarti <abhishek@taranjali.org>